backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pillow==10.1.0